requirements.txt
absl_py==2.0.0
anyascii==0.3.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
clearml==1.13.2
coqpit==0.0.17
cutlet==0.3.0
Cython==3.0.7
deepspeed==0.12.6
einops==0.7.0
encodec==0.1.1
faster_whisper==0.10.0
Flask==3.0.0
fsspec==2023.12.2
g2pkk==0.1.2
gradio==4.12.0
gruut==2.2.3
gruut_ipa==0.13.0
hangul_romanize==0.1.0
huggingface_hub==0.20.1
inflect==7.0.0
jamo==0.4.1
jieba==0.42.1
JPype1==1.5.0
k_diffusion==0.1.1.post1
kaggle==1.5.16
librosa==0.10.0
matplotlib==3.8.2
MeCab==0.996.3
mecab_python3==1.0.6
mutagen==1.47.0
num2words==0.5.13
numpy==1.22.0
onnxruntime==1.16.3
openai_whisper==20231117
opencc==1.1.7
packaging==23.2
pandas==1.5.3
pypinyin==0.50.0
pysbd==0.3.4
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.11.4
setuptools==63.2.0
six==1.16.0
soundfile==0.12.1
spacy==3.7.2
tokenizers==0.15.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
tqdm==4.66.1
transformers==4.36.2
umap_learn==0.5.5
Werkzeug==3.0.1
